+++ lib/widgets/surfaces/shared/tab_bar_bottom_internal.dart
@@ -704,6 +704,7 @@ class TabIndicatorState extends State<TabIndicator>
               quality: widget.quality,
               indicatorColor: indicatorColor,
               isBackgroundIndicator: false,
+              innerBlur: widget.innerBlur,
               borderRadius: thickness < 1 ? backgroundRadius : glassRadius,
               padding: const EdgeInsets.all(4),
               expansion: widget.indicatorExpansion,
@@ -783,6 +784,7 @@ class TabIndicatorState extends State<TabIndicator>
                     isBackgroundIndicator: false,
                     paintBackground: true,
                     paintGlass: false,
+                    innerBlur: widget.innerBlur,
                     borderRadius:
                         thickness < 1 ? backgroundRadius : glassRadius,
                     padding: const EdgeInsets.all(4),